# apps/backend-nest/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Vben Backend API",
    "description": "基于 NestJS + Zod 构建的现代化后端 API 服务\n\n## 架构特点\n- 使用 Zod 进行数据验证和类型定义\n- 统一的三段式返回格式 (code, message, data)\n- 自动生成的 OpenAPI 文档\n- 类型安全的 API 接口",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 认证令牌，格式：Bearer <token>"
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "post": {
        "summary": "创建用户",
        "description": "创建新用户账户，公共接口允许用户注册",
        "tags": [
          "用户管理"
        ],
        "operationId": "UserController_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "description": "用户名，必须唯一，用于登录认证"
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 100,
                    "format": "email",
                    "description": "用户邮箱地址，必须唯一，可用于找回密码等功能"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 255,
                    "description": "用户密码，存储加密后的密码哈希值"
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 50,
                    "description": "用户真实姓名"
                  },
                  "avatar": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "用户头像 URL，可选字段，用于显示用户头像"
                  },
                  "deptId": {
                    "type": "string",
                    "description": "所属部门ID"
                  },
                  "status": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1,
                    "default": 1,
                    "description": "用户账户状态：0-禁用，1-启用"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "响应状态码，200表示成功，其他表示错误"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息，描述操作结果"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "用户唯一标识符，使用 CUID 作为主键"
                        },
                        "username": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 50,
                          "description": "用户名，必须唯一，用于登录认证"
                        },
                        "email": {
                          "type": "string",
                          "maxLength": 100,
                          "format": "email",
                          "description": "用户邮箱地址，必须唯一，可用于找回密码等功能"
                        },
                        "password": {
                          "type": "string",
                          "maxLength": 255,
                          "description": "用户密码，存储加密后的密码哈希值"
                        },
                        "name": {
                          "type": "string",
                          "nullable": true,
                          "maxLength": 50,
                          "description": "用户真实姓名"
                        },
                        "avatar": {
                          "type": "string",
                          "nullable": true,
                          "maxLength": 500,
                          "description": "用户头像 URL，可选字段，用于显示用户头像"
                        },
                        "deptId": {
                          "type": "string",
                          "nullable": true,
                          "description": "所属部门ID"
                        },
                        "status": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1,
                          "default": 1,
                          "description": "用户账户状态：0-禁用，1-启用"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date",
                          "description": "用户创建时间，自动设置为当前时间"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date",
                          "description": "用户信息最后更新时间，每次更新时自动更新"
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "email",
                        "password",
                        "createdAt",
                        "updatedAt"
                      ],
                      "description": "响应数据"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "获取用户列表",
        "description": "支持分页查询，返回用户基本信息（不包含密码）",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "default": 1,
              "description": "页码，从1开始"
            },
            "required": false,
            "description": "页码，从1开始",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "description": "每页数量，最大100条"
            },
            "required": false,
            "description": "每页数量，最大100条",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按用户名筛选"
            },
            "required": false,
            "description": "按用户名筛选",
            "name": "username",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "email",
              "description": "按邮箱筛选"
            },
            "required": false,
            "description": "按邮箱筛选",
            "name": "email",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1,
              "description": "按账户状态筛选：0-禁用，1-启用"
            },
            "required": false,
            "description": "按账户状态筛选：0-禁用，1-启用",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按部门ID筛选"
            },
            "required": false,
            "description": "按部门ID筛选",
            "name": "deptId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "响应状态码，200表示成功，其他表示错误"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息，描述操作结果"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "用户唯一标识符，使用 CUID 作为主键"
                              },
                              "username": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 50,
                                "description": "用户名，必须唯一，用于登录认证"
                              },
                              "email": {
                                "type": "string",
                                "maxLength": 100,
                                "format": "email",
                                "description": "用户邮箱地址，必须唯一，可用于找回密码等功能"
                              },
                              "name": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 50,
                                "description": "用户真实姓名"
                              },
                              "avatar": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 500,
                                "description": "用户头像 URL，可选字段，用于显示用户头像"
                              },
                              "deptId": {
                                "type": "string",
                                "nullable": true,
                                "description": "所属部门ID"
                              },
                              "status": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1,
                                "default": 1,
                                "description": "用户账户状态：0-禁用，1-启用"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户创建时间，自动设置为当前时间"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户信息最后更新时间，每次更新时自动更新"
                              },
                              "department": {
                                "type": "object",
                                "nullable": true,
                                "properties": {
                                  "id": {
                                    "type": "string",
                                    "description": "部门唯一标识"
                                  },
                                  "name": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "description": "部门名称"
                                  },
                                  "pid": {
                                    "type": "string",
                                    "nullable": true,
                                    "description": "父级部门ID，支持树形结构"
                                  },
                                  "status": {
                                    "type": "integer",
                                    "minimum": 0,
                                    "maximum": 1,
                                    "description": "部门状态：0-禁用，1-启用"
                                  },
                                  "remark": {
                                    "type": "string",
                                    "nullable": true,
                                    "description": "备注信息"
                                  },
                                  "createTime": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "创建时间"
                                  },
                                  "updateTime": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "更新时间"
                                  }
                                },
                                "required": [
                                  "id",
                                  "name",
                                  "status",
                                  "createTime",
                                  "updateTime"
                                ],
                                "description": "所属部门信息"
                              },
                              "userRoles": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "string",
                                      "description": "关联记录唯一标识"
                                    },
                                    "userId": {
                                      "type": "string",
                                      "description": "用户ID"
                                    },
                                    "roleId": {
                                      "type": "string",
                                      "description": "角色ID"
                                    },
                                    "assignedAt": {
                                      "type": "string",
                                      "format": "date",
                                      "description": "分配时间"
                                    },
                                    "assignedBy": {
                                      "type": "string",
                                      "nullable": true,
                                      "description": "分配者ID（可选，记录是谁分配的角色）"
                                    },
                                    "role": {
                                      "type": "object",
                                      "properties": {
                                        "id": {
                                          "type": "string",
                                          "description": "角色唯一标识"
                                        },
                                        "name": {
                                          "type": "string",
                                          "maxLength": 100,
                                          "description": "角色名称（唯一）"
                                        },
                                        "status": {
                                          "type": "integer",
                                          "minimum": 0,
                                          "maximum": 1,
                                          "description": "角色状态：0-禁用，1-启用"
                                        },
                                        "remark": {
                                          "type": "string",
                                          "nullable": true,
                                          "description": "备注信息"
                                        },
                                        "createTime": {
                                          "type": "string",
                                          "format": "date",
                                          "description": "创建时间"
                                        },
                                        "updateTime": {
                                          "type": "string",
                                          "format": "date",
                                          "description": "更新时间"
                                        }
                                      },
                                      "required": [
                                        "id",
                                        "name",
                                        "status",
                                        "createTime",
                                        "updateTime"
                                      ],
                                      "description": "关联的角色信息"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "userId",
                                    "roleId",
                                    "assignedAt"
                                  ]
                                },
                                "description": "用户角色关联（多对多）"
                              }
                            },
                            "required": [
                              "id",
                              "username",
                              "email",
                              "createdAt",
                              "updatedAt"
                            ]
                          },
                          "description": "数据列表"
                        },
                        "total": {
                          "type": "number",
                          "description": "总记录数"
                        },
                        "page": {
                          "type": "number",
                          "description": "当前页码"
                        },
                        "limit": {
                          "type": "number",
                          "description": "每页数量"
                        },
                        "totalPages": {
                          "type": "number",
                          "description": "总页数"
                        }
                      },
                      "required": [
                        "items",
                        "total",
                        "page",
                        "limit",
                        "totalPages"
                      ],
                      "description": "响应数据"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "未授权访问"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}